+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `Innmind\Immutable\Sequence::windows()`
+
 ## 5.16.0 - 2025-06-02
 
 ### Added